+++ server/lib/u.js
@@ -1,0 +1,5 @@
+import url from 'url';
+
+let q = url.parse('/data/#hum?q=1',true)
+
+console.log(q)