+++ server/main.js
@@ -1,20 +1,48 @@
-import sirv from 'sirv';
 import express from 'express';   // dont use polka, does not send SSE messages
+import * as dotenv from 'dotenv' // see https://github.com/motdotla/dotenv#how-do-i-use-dotenv-with-import
+import helmet from "helmet";
 import compression from 'compression';
+import sirv from 'sirv';
+import cors from 'cors';
+//import morgan from 'morgan';  // logger, maybe winston ?
 
 import dynamicPages from './dynamicPages.js';
+import redirectToSlash from '#root/server/lib/redirectToSlash.js';
 
+import {dbConnect} from './db/sqlite/connect.js';
+
+dotenv.config()
 const { PORT, NODE_ENV } = process.env;
 const dev = NODE_ENV === 'development';
 const port = parseInt(PORT || '3000');
 
-
+const db = dbConnect();
 
 console.log(`listening in http://localhost:${port}`);
 
-express() 
+const srv = express() 
+
+// -- BOILERLATE
+
+// -- behind reverse proxy ?
+// app.set('trust proxy', '127.0.0.1');
+// app.use(morgan('dev'))
+srv.use(helmet());
+// srv.use(cors())
+srv.use(express.json())
+srv.use(express.urlencoded({extended: false}))
+// status check
+srv.get('/status', (req,res)=> res.sendStatus(200).end());	
+
+// -- debug
+// srv.use((req,res,next)=>{
+// 	console.log('url:', req.path, req.query)
+// 	next()
+// })
+
+
   // SSE
-	.get('/sse', function(req, res) {
+srv.get('/sse', function(req, res) {
 		res.writeHead(200, {
       'Content-Type': 'text/event-stream',
       'Connection': 'keep-alive',
@@ -30,13 +58,14 @@ express()
 		
 	})
 
-	.use(
-		compression({ threshold: 0 }),
-		sirv('static', { dev }),
-    dynamicPages
+srv.use(
+	  redirectToSlash,
+		// compression({ threshold: 0 }),
+    dynamicPages,
+		sirv('server/pages', { dev })
 	)
 
-	.listen(port, err => {
+srv.listen(port, err => {
 		if (err) console.log('error', err);
 	});