+++ server/pages/js/debug.js
@@ -1,7 +1,7 @@
 document.addEventListener('dblclick',ev=>{
   // -- reload whole page
-  // location.reload()
+  location.reload()
 
   // -- or just reload css
-  document.querySelector('head > link').href = '/style.css?' + Date.now()
+  //document.querySelector('head > link').href = '/style.css?' + Date.now()
 });