+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 
 # TODO
-- login
-- sqlite
+- post, fetchJson
 - colocated queries client/server sharing types
-
+- login
+- aclarar server/client comp (cascade)
 
 # DONE
+- sqlite
+- trailing slashs redirec, so import 'index.js' work properly in client
+- removed static/ now everything goes on pages (dynamicPages prevents downloading of server code)
 - layout
 - debug: lajout.js, on doubleclick reload page OR just css
 - added jsconfig.json for typecheck